fix building if CC is exported
the makefile sets CC=valac. If CC is exported, then this overrides the exported value of CC
make calls CC=valac. valac sees that CC is exported and wants to honour it, calling valac as the C compiler, which fails.

check by adding --verbose after $(CC) in the Makefile and exporting e.g. CC=gcc

change the Makefile variable from CC to VALAC, which makes more sense
anyway, as valac isn't a C compiler but a Vala compiler